+++ src/components/ui-checkbox.vue
@@ -22,7 +22,7 @@
     </div>
 </template>
 <script>
-import UiLabel from '@/components/ui-label.vue';
+import UiLabel from './ui-label.vue';
 
 export default {
     components: {
@@ -36,6 +36,7 @@ export default {
         },
         selected: {
             type: String,
+            default: '',
         },
     },
 };