+++ src/components/ui-select.vue
@@ -1,7 +1,7 @@
 <template>
     <div class="relative">
         <select
-            class="appearance-none cursor-pointer w-full rounded-sm border bg-white shadow-sm py-2 px-4 outline-none text-base font-normal"
+            class="appearance-none cursor-pointer w-full rounded-sm border bg-white shadow-sm py-2 px-4 outline-none text-base font-normal h-11"
             :class="{
                 'border-light-1 focus:border-dark-3': isValid,
                 'border-red-1 focus:border-red-1': !isValid,